@@ -18,7 +18,7 @@ class Exponential(RandomVariable):
     lambda_: float = 1
 
     def __str__(self) -> str:
-        return "Exponential(λ={})".format(self.lambda_)
+        return f"Exponential(λ={self.lambda_})"
 
     def pdf(self, x: float) -> float:
         return 0
